pyproject.toml
[project]
name = "fun4kids"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.108.1",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
]
